+++ precompiles/precompile.go
@@ -565,6 +565,8 @@ func Precompiles() map[addr]ArbosPrecompile {
 	ArbOwnerPublic.methodsByName["GetMaxStylusContractFragments"].arbosVersion = params.ArbosVersion_StylusContractLimit
 	ArbOwnerPublic.methodsByName["GetDeriwRouterConfig"].arbosVersion = params.ArbosVersion_60
 	ArbOwnerPublic.methodsByName["GetScheduledDeriwRouterConfig"].arbosVersion = params.ArbosVersion_60
+	ArbOwnerPublic.methodsByName["GetDeriwOSVersion"].arbosVersion = params.ArbosVersion_60
+	ArbOwnerPublic.methodsByName["GetScheduledDeriwOSUpgrade"].arbosVersion = params.ArbosVersion_60
 
 	ArbWasmImpl := &ArbWasm{Address: types.ArbWasmAddress}
 	ArbWasm := insert(MakePrecompile(precompilesgen.ArbWasmMetaData, ArbWasmImpl))
@@ -635,6 +637,8 @@ func Precompiles() map[addr]ArbosPrecompile {
 	ArbOwner.methodsByName["SetCollectTips"].arbosVersion = params.ArbosVersion_60
 	ArbOwner.methodsByName["ScheduleDeriwRouterConfig"].arbosVersion = params.ArbosVersion_60
 	ArbOwner.methodsByName["CancelScheduledDeriwRouterConfig"].arbosVersion = params.ArbosVersion_60
+	ArbOwner.methodsByName["ScheduleDeriwOSUpgrade"].arbosVersion = params.ArbosVersion_60
+	ArbOwner.methodsByName["CancelScheduledDeriwOSUpgrade"].arbosVersion = params.ArbosVersion_60
 	stylusMethods := []string{
 		"SetInkPrice", "SetWasmMaxStackDepth", "SetWasmFreePages", "SetWasmPageGas",
 		"SetWasmPageLimit", "SetWasmMinInitGas", "SetWasmInitCostScalar",