+++ precompiles/precompile_test.go
@@ -193,7 +193,7 @@ func TestPrecompilesPerArbosVersion(t *testing.T) {
 		params.ArbosVersion_41: 10,
 		params.ArbosVersion_50: 9,
 		params.ArbosVersion_59: 2,
-		params.ArbosVersion_60: 26,
+		params.ArbosVersion_60: 30,
 	}
 
 	precompiles := Precompiles()